// frontend/src/codegen/model/curatedAppSelection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Flathub API
 * OpenAPI spec version: 0.1.0
 */
import type { CuratedAppSelectionApp } from "./curatedAppSelectionApp"

export interface CuratedAppSelection {
  id: number
  theme_key: string
  slot: string
  starts_at: string
  ends_at: string
  apps: CuratedAppSelectionApp[]
}